Entfernt DownloadEntityState für einen allgemeinen FileStatusEntityState

+++ src/app/services/file-watcher.service.ts
@@ -1,20 +1,15 @@
 import { Injectable, OnDestroy } from '@angular/core';
 import { FirebaseService } from './firebase.service';
-import { from, Observable, of, Subject, Subscription } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 import { Action } from '@ngrx/store';
 import { FileRequest } from '../models/file-request.model';
 import { SharedFile } from '../models/shared-file.model';
-import { catchError, flatMap, map } from 'rxjs/operators';
 import { HttpEventType } from '@angular/common/http';
-import {
-  DownloadActionTypes,
-  DownloadFailed,
-  DownloadFinished,
-  DownloadProgressUpdate,
-  DownloadStarted,
-} from '../actions/download.actions';
 import { MatSnackBar } from '@angular/material';
 import { AddFileToFileRequest } from '../actions/file-request.actions';
+import { UpsertFileStatus } from '../actions/file-status.actions';
+import { FileStatus, FileStatusType } from '../models/file-status.model';
+import { UpdateSharedFile } from '../actions/shared-file.actions';
 
 @Injectable({
   providedIn: 'root',
@@ -55,24 +50,47 @@ export class FileWatcherService implements OnDestroy {
           const sharedFile: SharedFile = { ...fileAction.payload.sharedFile, fileRequest: fileRequest.id };
           if (sharedFile.uploadedAt) {
             // Downloads the file
-            this.downloadFile(sharedFile).subscribe((downloadAction) => {
-              switch (downloadAction.type) {
-                // Checks if the file was downloaded successfully
-                case DownloadActionTypes.DownloadFinished:
-                  // Deletes the file from the server
-                  this.firebaseService.deleteFile(sharedFile).subscribe({
-                    // Ignores errors or a successful deletion, since a cloud function cleans up old files anyway
-                    error: () => {
-                    },
-                  });
+            this.firebaseService.downloadFile(sharedFile).subscribe((event) => {
+              let fileStatus = <FileStatus>{ id: sharedFile.id };
+              switch (event.type) {
+                case HttpEventType.Sent:
+                  fileStatus = { ...fileStatus, type: FileStatusType.DownloadStarted, message: 'Herunterladen...' };
                   break;
-                // Checks if the download failed and displays a snackbar
-                case DownloadActionTypes.DownloadFailed:
-                  this.snackbar.open(`Fehler beim Laden von '${ sharedFile.fileName }' ${ downloadAction.payload.error.message }`);
+                case HttpEventType.DownloadProgress: {
+                  const loaded = event.loaded;
+                  const total = event.total;
+                  const bytes = { loaded, total };
+                  fileStatus = { ...fileStatus, type: FileStatusType.DownloadUpdate, message: 'Herunterladen...', bytes };
                   break;
+                }
+                case HttpEventType.Response: {
+                  if (event.ok) { // Checks if the file was downloaded successfully
+                    const totalBytes = event.body ? event.body.size : 0;
+                    const bytes = { loaded: totalBytes, total: totalBytes };
+
+                    fileStatus = { ...fileStatus, type: FileStatusType.DownloadCompleted, message: 'Download abgeschlossen', bytes };
+
+                    // Updates the shared file; important that this happens before the file status action is emitted
+                    this.actionsSubject.next(new UpdateSharedFile({ sharedFile: { id: sharedFile.id, changes: { blob: event.body } } }));
+
+                    // Deletes the file from the server
+                    this.firebaseService.deleteFile(sharedFile).subscribe({
+                      // Ignores errors or a successful deletion, since a cloud function cleans up old files anyway
+                      error: () => {
+                      },
+                    });
+                  } else {
+                    // Emits download failed
+                    // displays a snackbar // this.snackbar.open(`Fehler beim Laden von '${ sharedFile.fileName }' ${ event.statusText }`);
+                    fileStatus = { ...fileStatus, message: 'Download fehlgeschlagen', type: FileStatusType.Error };
+                  }
+                  break;
+                }
               }
               // Emits the action about the download status of the file
-              this.actionsSubject.next(downloadAction);
+              if (fileStatus) {
+                this.updateFileStatus(fileStatus);
+              }
             });
           }
           // Emits the file action
@@ -97,7 +115,11 @@ export class FileWatcherService implements OnDestroy {
     }
   }
 
-  downloadFile(sharedFile: SharedFile): Observable<DownloadStarted | DownloadProgressUpdate | DownloadFinished | DownloadFailed> {
+  updateFileStatus(fileStatus: FileStatus) {
+    this.actionsSubject.next(new UpsertFileStatus({ fileStatus }));
+  }
+
+  /*downloadFile(sharedFile: SharedFile): Observable<DownloadStarted | DownloadProgressUpdate | DownloadFinished | DownloadFailed> {
     return this.firebaseService.downloadFile(sharedFile).pipe(
       flatMap((event) => {
         switch (event.type) {
@@ -127,6 +149,6 @@ export class FileWatcherService implements OnDestroy {
       }),
       catchError((error) => of(new DownloadFailed({ sharedFileId: sharedFile.id, error }))),
     );
-  }
+  }*/
 
 }
